[FIX] ItemData ScriptableObject 문제 수정
