+++ Archero/Assets/New Armor Item Data.asset
@@ -9,9 +9,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: ArmorItemData
-  m_EditorClassIdentifier: Assembly-CSharp::ArmorItemData
+  m_Script: {fileID: 11500000, guid: 4f2ce75b7bb62a742bd96b8455ab13bc, type: 3}
+  m_Name: New Armor Item Data
+  m_EditorClassIdentifier: 
   Name: "\uC544\uBA38"
   Icon: {fileID: 0}
   Animator: {fileID: 0}
