+++ Archero/Assets/New Armor Item Data.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 63677da8e71c4da4fba4bde2b5b4a17e
+guid: a815db93a5f4ea64abace585b4da9235
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
