+++ Archero/Assets/New EXP Item Data.asset
@@ -9,9 +9,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: EXPItemData
-  m_EditorClassIdentifier: Assembly-CSharp::EXPItemData
+  m_Script: {fileID: 11500000, guid: 8f427cbe6cbcfca4dba1a1c8df1a8e90, type: 3}
+  m_Name: New EXP Item Data
+  m_EditorClassIdentifier: 
   Name: "\uC774\uC5D1\uC2A4\uD53C"
   Icon: {fileID: 0}
   Animator: {fileID: 0}
