+++ Archero/Assets/New EXP Item Data.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 119ecb7f1d545c04eacf0260c25b2377
+guid: 0447e14641601634f839bf1b5ef039c2
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
