+++ Archero/Assets/New Potion.asset
@@ -9,14 +9,11 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: PotionItemData
-  m_EditorClassIdentifier: Assembly-CSharp::PotionItemData
-  Name: "\uD3EC\uC1581"
+  m_Script: {fileID: 11500000, guid: 02f1355be0710b54ea2ed8580c2a6745, type: 3}
+  m_Name: New Potion
+  m_EditorClassIdentifier: 
+  Name: "\uD3EC\uC158"
   Icon: {fileID: 0}
   Animator: {fileID: 0}
   Type: 0
-  OnCollected:
-    m_PersistentCalls:
-      m_Calls: []
   HealAmount: 0
