+++ Archero/Assets/New Potion.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a5a374b100b91f94ead8465bc9a2687a
+guid: 087afaf7eda10c84e817393ccaa59080
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
