+++ Archero/Assets/New Weapon Item Data.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 60a507da28f556743af4fafbeb0170dc
+guid: 58ac0240a8cba544499b6de2a74ca19e
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
