+++ Archero/Assets/Scripts/Item/ArmorItemData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4f2ce75b7bb62a742bd96b8455ab13bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
