+++ Archero/Assets/Scripts/Item/EXPItemData.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(menuName = "ScriptableObjects/ItemData/Collectable/EXP")]
+public class EXPItemData : CollectableItemData
+{
+    public int EXPAmount; // È¸º¹·®
+}
