+++ Archero/Assets/Scripts/Item/EXPItemData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f427cbe6cbcfca4dba1a1c8df1a8e90
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
