+++ Archero/Assets/Scripts/Item/ItemData.cs
@@ -31,56 +31,8 @@ public abstract class CollectableItemData : BaseItemData
     //public UnityEvent<Player> OnCollected;
 }
 
-[CreateAssetMenu(menuName = "ScriptableObjects/ItemData/Collectable/Potion")]
-public class PotionItemData : CollectableItemData
-{
-    public int HealAmount;
-}
-
-[CreateAssetMenu(menuName = "ScriptableObjects/ItemData/Collectable/EXP")]
-public class EXPItemData : CollectableItemData
-{
-    public int EXPAmount; // 회복량
-}
-
 public abstract class EquippableItemData : BaseItemData
 {
     public abstract void Equip();
     public abstract void UnEquip();
-}
-
-
-[CreateAssetMenu(menuName = "ScriptableObjects/ItemData/Equippable/Weapon")]
-public class WeaponItemData : EquippableItemData
-{
-    public int AttackPower; // 공격력
-    public override void Equip()
-    {
-    }
-    public override void UnEquip()
-    {
-    }
-    //public override void Equip(Player player)
-    //{
-    //}
-    //public override void UnEquip(Player player)
-    //{
-    //}
-}
-[CreateAssetMenu(menuName = "ScriptableObjects/ItemData/Equippable/Armor")]
-public class ArmorItemData : EquippableItemData
-{
-    public int DefensePower; // 공격력
-    public override void Equip()
-    {
-    }
-    public override void UnEquip()
-    {
-    }
-    //public override void Equip(Player player)
-    //{
-    //}
-    //public override void UnEquip(Player player)
-    //{
-    //}
 }
