+++ Archero/Assets/Scripts/Item/Potion.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(menuName = "ScriptableObjects/ItemData/Collectable/Potion")]
+public class Potion : BaseItemData
+{
+    public int HealAmount;
+}
