+++ Archero/Assets/Scripts/Item/Potion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 02f1355be0710b54ea2ed8580c2a6745
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
