+++ Archero/Assets/Scripts/Item/WeaponItemData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 99bbd6d2c8751944dab928d17af2007c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
